+++ util/opentelemetry-util-http/src/opentelemetry/util/http/__init__.py
@@ -15,6 +15,15 @@
 from os import environ
 from re import compile as re_compile
 from re import search
+from typing import Callable, Collection, Optional, Tuple
+
+from opentelemetry import trace
+
+
+_root = r"OTEL_PYTHON_{}"
+
+HTTP_HEADER_ACCESS_CONTROL_EXPOSE_HEADERS = "Access-Control-Expose-Headers"
+ENV_HTTP_TRACE_RESPONSE_HEADER = "OTEL_PYTHON_HTTP_TRACE_RESPONSE_HEADER"
 
 
 class ExcludeList:
@@ -27,9 +36,6 @@ class ExcludeList:
 
     def url_disabled(self, url: str) -> bool:
         return bool(self._excluded_urls and search(self._regex, url))
-
-
-_root = r"OTEL_PYTHON_{}"
 
 
 def get_traced_request_attrs(instrumentation):
@@ -57,3 +63,30 @@ def get_excluded_urls(instrumentation):
         ]
 
     return ExcludeList(excluded_urls)
+
+
+def get_trace_response_headers(
+    span: trace.Span,
+) -> Collection[Tuple[str, str]]:
+    header_name = environ.get(ENV_HTTP_TRACE_RESPONSE_HEADER, "").strip()
+    if not header_name:
+        return tuple()
+
+    if span is trace.INVALID_SPAN:
+        return tuple()
+
+    ctx = span.get_span_context()
+    if ctx is trace.INVALID_SPAN_CONTEXT:
+        return tuple()
+
+    return (
+        (HTTP_HEADER_ACCESS_CONTROL_EXPOSE_HEADERS, header_name),
+        (
+            header_name,
+            'traceparent;desc="00-{trace_id}-{span_id}-{sampled}"'.format(
+                trace_id=trace.format_trace_id(ctx.trace_id),
+                span_id=trace.format_span_id(ctx.span_id),
+                sampled="01" if ctx.trace_flags.sampled else "00",
+            ),
+        ),
+    )